+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 - warning about outdated metadata, in case an inclusion/exclusion pattern does not match any files in an Asset anymore.
+- support for rendering package identifiers in metadata description text, using syntax `` `pkg=group:name` ``.
 
 
 ## [0.4.2] - 2024-04-18